DIS-1076: Add back some old deletion logic for Portal/Custom Pages.

+++ code/web/sys/WebBuilder/PortalPage.php
@@ -241,8 +241,10 @@ class PortalPage extends DB_LibraryLinkedObject {
 			$this->clearAudiences();
 			$this->clearCategories();
 			$this->clearAccess();
-			foreach ($this->getRows() as $row) {
-				$row->delete();
+			if (!$useWhere) {
+				foreach ($this->getRows() as $row) {
+					$row->delete();
+				}
 			}
 		}
 		return $ret;
